% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AddunEvaled}
\alias{AddunEvaled}
\title{Title}
\usage{
AddunEvaled(key, xdata, xdatar, xdatab, xdatah, UnEvaledlist)
}
\arguments{
\item{key}{The key values of the evaluated models}

\item{xdata}{Intervention variables}

\item{xdatar}{Intervention impact model parameter r}

\item{xdatab}{Intervention impact model parameter b}

\item{xdatah}{Intervention impact model parameter h}

\item{UnEvaledlist}{List of unassessed models}
}
\value{
Updated list of unassessed models.
}
\description{
Title
}
