% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ArimaOptExtr}
\alias{ArimaOptExtr}
\title{Extract the p, d, and q parameters of the ARIMA model.}
\usage{
ArimaOptExtr(modelAR)
}
\arguments{
\item{modelAR}{A fitted ARIMA model.}
}
\value{
The ArimaOptExtr function returns a vector which is the p,d,q parameters
}
\description{
Extract the p, d, and q parameters of the ARIMA model.
}
