% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_middle_position}
\alias{find_middle_position}
\title{Find the intermediate position corresponding to the maximum positive or negative change in the v-weights of the LTF model}
\usage{
find_middle_position(x)
}
\arguments{
\item{x}{LTF model v-weights}
}
\value{
the intermediate position corresponding to the maximum positive or negative change
}
\description{
Find the intermediate position corresponding to the maximum positive or negative change in the v-weights of the LTF model
}
