% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FindMinAic}
\alias{FindMinAic}
\title{Find the optimal model with the smallest AIC from the evaluated models}
\usage{
FindMinAic(Evaled)
}
\arguments{
\item{Evaled}{the evaluated models}
}
\value{
The optimal model's name and its AIC value.
}
\description{
Find the optimal model with the smallest AIC from the evaluated models
}
