% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{IdenPatternb}
\alias{IdenPatternb}
\title{Using the identification rules of the LTF model, automatically identify possible sets of b}
\usage{
IdenPatternb(modelxp)
}
\arguments{
\item{modelxp}{a fitted LTF model}
}
\value{
the possible sets of b
}
\description{
Using the identification rules of the LTF model, automatically identify possible sets of b
}
