% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LTFxpFun}
\alias{LTFxpFun}
\title{Building the LTF model}
\usage{
LTFxpFun(y, x_p, rlist, LTFmaxk)
}
\arguments{
\item{y}{a univariate time series}

\item{x_p}{a pulse function interrupted by an intervention event}

\item{rlist}{The ARIMA model order (p, d, q) of noise component}

\item{LTFmaxk}{The longest time-lagged response k in the LTF model}
}
\value{
a fitted LTF model
}
\description{
Building the LTF model
}
