% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Optarimax}
\alias{Optarimax}
\title{Fitting an ARIMAx model with known initial values for the intervention component and the intercept term}
\usage{
Optarimax(y, order, xtransf, transfer, inintra)
}
\arguments{
\item{y}{a univariate time series}

\item{order}{regular ARIMA order}

\item{xtransf}{intervention variables}

\item{transfer}{a list consisting of the ARMA orders for each transfer (distributed lag) covariate}

\item{inintra}{initial values for the intervention component and the intercept term}
}
\value{
The Optarimax function returns the ARIMAX model fitted with known initial values for the intervention component and the intercept term
}
\description{
Fitting an ARIMAx model with known initial values for the intervention component and the intercept term
}
