% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Tracoef}
\alias{Tracoef}
\title{Extract the intervention component and intercept coefficients of the ARIMA model.}
\usage{
Tracoef(mode)
}
\arguments{
\item{mode}{a fitted LTF model}
}
\value{
the coefficients of the intervention component and intercept
}
\description{
Extract the intervention component and intercept coefficients of the ARIMA model.
}
